Add requirements.txt file
The packages versions in `requirements.txt` were the most recently
released versions when the data was presented today (2021-09-10). The
objective is to limit issues related to code rot and maintain
consistency between the version of `pandas` used when pickling data and
the version of `pandas` used when unpickling and using data.